refactor: moves trade history cubit to coin directory
