+++ lib/features/coin/presentation/cubits/trade_history_cubit/trade_history_state.dart
