+++ lib/injection_container.dart
@@ -16,7 +16,7 @@ import 'package:polkadex/common/orderbook/domain/repositories/iorderbook_reposit
 import 'package:polkadex/common/orderbook/domain/usecases/fetch_orderbook_data_usecase.dart';
 import 'package:polkadex/common/orderbook/presentation/cubit/orderbook_cubit.dart';
 import 'package:polkadex/common/trades/domain/usecases/get_trades_usecase.dart';
-import 'package:polkadex/common/trades/presentation/cubits/trade_history_cubit/trade_history_cubit.dart';
+import 'package:polkadex/features/coin/presentation/cubits/trade_history_cubit/trade_history_cubit.dart';
 import 'package:polkadex/features/coin/data/datasources/coin_remote_datasource.dart';
 import 'package:polkadex/features/coin/data/repositories/coin_repository.dart';
 import 'package:polkadex/features/coin/domain/repositories/icoin_repository.dart';
@@ -62,7 +62,7 @@ import 'common/orderbook/data/repositories/orderbook_repository.dart';
 import 'common/orderbook/domain/usecases/fetch_orderbook_live_data_usecase.dart';
 import 'common/trades/domain/usecases/get_orders_usecase.dart';
 import 'features/coin/domain/usecases/withdraw_usecase.dart';
-import 'features/coin/presentation/cubits/withdraw_cubit.dart';
+import 'features/coin/presentation/cubits/withdraw_cubit/withdraw_cubit.dart';
 import 'features/setup/data/datasources/account_local_datasource.dart';
 import 'common/cubits/account_cubit/account_cubit.dart';
 import 'features/setup/domain/repositories/imnemonic_repository.dart';
